refactor: 2단계 피드백 반영
- 이메일 유효성 검증 어노테이션 수정.
- JwtUtil 클래스 추가.
