+++ src/main/java/gift/config/FilterConfig.java
@@ -2,6 +2,7 @@ package gift.config;
 
 import gift.filter.JwtAuthenticationFilter;
 
+import gift.util.JwtUtil;
 import jakarta.servlet.Filter;
 import org.springframework.boot.web.servlet.FilterRegistrationBean;
 import org.springframework.context.annotation.Bean;
@@ -10,10 +11,16 @@ import org.springframework.context.annotation.Configuration;
 @Configuration
 public class FilterConfig {
 
+    private final JwtUtil jwtUtil;
+
+    public FilterConfig(JwtUtil jwtUtil) {
+        this.jwtUtil = jwtUtil;
+    }
+
     @Bean
     public FilterRegistrationBean<Filter> jwtFilter() {
         FilterRegistrationBean<Filter> registration = new FilterRegistrationBean<>();
-        registration.setFilter(new JwtAuthenticationFilter());
+        registration.setFilter(new JwtAuthenticationFilter(jwtUtil));
         registration.addUrlPatterns("/*"); // 전체 경로 감시
         registration.setOrder(1);
         return registration;
