+++ src/main/java/gift/dto/LoginMemberDto.java
@@ -1,5 +1,6 @@
 package gift.dto;
 
+import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 
 public class LoginMemberDto {
@@ -8,6 +9,7 @@ public class LoginMemberDto {
     private Long id;
 
     @NotBlank(message = "이메일는 필수 입력 값입니다.")
+    @Email(message = "유효한 이메일 형식이 아닙니다.")
     private String email;
 
     public LoginMemberDto(String email) {
