+++ src/main/java/gift/dto/RegisterRequest.java
@@ -1,12 +1,13 @@
 package gift.dto;
 
+import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
 
 public class RegisterRequest {
 
     @NotBlank(message = "이메일는 필수 입력 값입니다.")
-    @Pattern(regexp = "^[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\\.[A-Za-z]{2,}$", message = "유효한 이메일 형식이 아닙니다.")
+    @Email(message = "유효한 이메일 형식이 아닙니다.")
     private String email;
 
     @NotBlank(message = "비밀번호는 필수 입력 값입니다.")
